modified:   ../AStar/AStar.cpp 	modified:   ../AStar/AStar.hpp

+++ GraphSearchAlgorithms/AStar/AStar.hpp
@@ -18,7 +18,7 @@ class AStar : public graphList{
     
 public:
     AStar(int num_vertices) : graphList(num_vertices){} 
-    int heuristic(Node& start, Node& goal); //most likely euclidean distances times weight 
+    double heuristic(Node& start, Node& goal); //most likely euclidean distances times weight 
     int solution_astar(char src, char dest); 
 
 };
